backend/src/requirements.txt
fuzzywuzzy
gensim
nltk
numpy
python-Levenshtein
pyyaml
sklearn